requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
beautifulsoup4
SQLAlchemy
psycopg2-binary
APScheduler
python-telegram-bot==20.3
httpx
python-multipart